Добавил AlertDialog, Toast и Snackbar

+++ app/src/main/java/com/example/gb03_android_on_java_notes/ui/list/ListFragment.java
@@ -1,10 +1,14 @@
 package com.example.gb03_android_on_java_notes.ui.list;
 
+import static com.example.gb03_android_on_java_notes.utils.NoteUtils.colorCircleDrawable;
+
+import android.app.AlertDialog;
 import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -16,6 +20,8 @@ import com.example.gb03_android_on_java_notes.App;
 import com.example.gb03_android_on_java_notes.R;
 import com.example.gb03_android_on_java_notes.domain.Note;
 import com.example.gb03_android_on_java_notes.domain.NoteRepository;
+import com.google.android.material.snackbar.BaseTransientBottomBar;
+import com.google.android.material.snackbar.Snackbar;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -27,6 +33,7 @@ public class ListFragment extends Fragment implements NoteViewHolder.Callbacks {
     private NoteAdapter adapter;
     private Context context;
     private Controller controller;
+    private RecyclerView recyclerView;
 
     public static ListFragment getInstance() {
         return new ListFragment();
@@ -62,10 +69,10 @@ public class ListFragment extends Fragment implements NoteViewHolder.Callbacks {
     }
 
     private void initRecycler(View view) {
-        RecyclerView noteRecyclerView = view.findViewById(R.id.note_recycler_view);
-        noteRecyclerView.setLayoutManager(new LinearLayoutManager(context));
+        recyclerView = view.findViewById(R.id.note_recycler_view);
+        recyclerView.setLayoutManager(new LinearLayoutManager(context));
         adapter = new NoteAdapter(repository.getNotes(), this);
-        noteRecyclerView.setAdapter(adapter);
+        recyclerView.setAdapter(adapter);
     }
 
     @Override
@@ -80,19 +87,36 @@ public class ListFragment extends Fragment implements NoteViewHolder.Callbacks {
 
     @Override
     public boolean onLongClickItem(Note note) {
-        if (repository.removeNote(note.getId())) {
-            int position = positionOf(note);
-            adapter.notifyItemRemoved(position);
-            adapter.notifyItemRangeChanged(position, adapter.getItemCount() - position);
-            controller.onNoteDeleted(note);
-            return true;
-        }
-        return false;
+        showDeleteNoteDialog(note);
+        return true;
+    }
+
+    private void showDeleteNoteDialog(Note note) {
+        new AlertDialog.Builder(context)
+                .setTitle(note.getHeader())
+                .setMessage(R.string.ask_delete)
+                .setIcon(colorCircleDrawable(context, note.getColor()))
+                .setPositiveButton(R.string.yes, (dialog, which) -> {
+                    if (repository.removeNote(note.getId())) {
+                        int position = positionOf(note);
+                        adapter.notifyItemRemoved(position);
+                        adapter.notifyItemRangeChanged(position, adapter.getItemCount() - position);
+                        Toast.makeText(context, R.string.note_deleted, Toast.LENGTH_SHORT).show();
+                        controller.onNoteDeleted(note);
+                    }
+                })
+                .setNegativeButton(R.string.no, null)
+                .show();
     }
 
     private void onClickAddNoteButton(View view) {
         Note note = repository.createNote();
         controller.onNoteSelected(note);
+
+        Snackbar.make(recyclerView, R.string.note_created, BaseTransientBottomBar.LENGTH_LONG)
+                .setAnimationMode(BaseTransientBottomBar.ANIMATION_MODE_SLIDE)
+                .setBackgroundTint(context.getColor(R.color.peach))
+                .show();
     }
 
     private int positionOf(Note note) {
@@ -107,6 +131,7 @@ public class ListFragment extends Fragment implements NoteViewHolder.Callbacks {
 
     public interface Controller {
         void onNoteSelected(Note note);
+
         void onNoteDeleted(Note note);
     }
